change to use npm instead of yarn, clean up layers, use node 20
