+++ arSam/layers/constantsLayer/nodejs/package.json
@@ -1,6 +1,6 @@
 {
-    "name": "ConstantsLayer",
-    "version": "0.0.1",
-    "author": "Digitalspace <contact@digitalspace.ca>",
-    "license": "MIT"
-  } 
+  "name": "ConstantsLayer",
+  "version": "0.0.1",
+  "author": "Digitalspace <contact@digitalspace.ca>",
+  "license": "MIT"
+}
