+++ arSam/layers/functionsLayer/Makefile
@@ -4,7 +4,7 @@ build-FunctionsLayer:
 					# remove makefile in artifacts directory
 					rm "$(ARTIFACTS_DIR)/Makefile"
 					# install dependencies in artifacts directory
-					cd nodejs && yarn cache clean && yarn --update-checksums && yarn install --ignore-scripts --frozen-lockfile --modules-folder "$(ARTIFACTS_DIR)/nodejs/node_modules"
+					cd nodejs && npm install --prefix "$(ARTIFACTS_DIR)"/nodejs
 					# symlink to node_modules folder for testing purposes
 					cd "$(ARTIFACTS_DIR)" && ln -s "$(ARTIFACTS_DIR)"/nodejs/node_modules node_modules
 					# remove package.json to avoid rebuilding when changes don't relate to dependencies
