--- arSam/layers/varianceLayer/nodejs/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "varianceLayer",
-  "version": "0.0.1",
-  "author": "Digitalspace <contact@digitalspace.ca>",
-  "license": "MIT"
-}
